Refactor geolocation api related code
