+++ src/constants/constants.ts
@@ -1,3 +1,7 @@
-export const GEOLOCATION_API_URL = process.env.NEXT_PUBLIC_GEOLOCATION_API_ENDPOINT || ''
+export const GEOLOCATION_API_ENDPOINT = process.env.NEXT_PUBLIC_GEOLOCATION_API_ENDPOINT || ''
 export const GEOLOCATION_API_KEY = process.env.GEOLOCATION_API_KEY || ''
 export const GEOLOCATION_CACHE_LIFETIME = Number(process.env.NEXT_PUBLIC_GEOLOCATION_CACHE_LIFETIME) || 0
+
+export const WEATHER_API_BASE_ENDPOINT = process.env.NEXT_PUBLIC_WEATHER_API_BASE_ENDPOINT || ''
+export const GEOCODING_API_ENDPOINT = process.env.NEXT_PUBLIC_GEOCODING_API_ENDPOINT || ''
+export const OPEN_WEATHER_API_KEY = process.env.OPEN_WEATHER_API_KEY || ''
