+++ src/services/queries/getUserLocation.ts
@@ -1,15 +1,21 @@
 // types
-import { GeolocationApiResponse } from '@/types/geolocation'
+import { GeolocationApiResponse } from '@/types/geolocationApi'
 
 // constants
 import { GEOLOCATION_CACHE_LIFETIME } from '@/constants/constants'
 
-// util
-import { getGeolocationApiUrl } from '@/utils/geolocation'
+// utils
+import { getGeolocationApiEndpoint } from '@/utils/geolocation'
 
+/**
+ * Geolocation API call
+ *
+ * @param {string[]} [fields]
+ * @return {*}  {Promise<Partial<GeolocationApiResponse>>}
+ */
 export const getUserLocation = async (fields?: string[]): Promise<Partial<GeolocationApiResponse>> => {
   const res = await fetch(
-    getGeolocationApiUrl(fields),
+    getGeolocationApiEndpoint(fields),
     {
       cache: 'no-store',
       next: {
