+++ src/utils/geolocation.ts
@@ -1,4 +1,4 @@
-import { GEOLOCATION_API_KEY, GEOLOCATION_API_URL } from '@/constants/constants'
+import { GEOLOCATION_API_KEY, GEOLOCATION_API_ENDPOINT } from '@/constants/constants'
 
 /**
  * Returns geolocation API url with params
@@ -6,7 +6,7 @@ import { GEOLOCATION_API_KEY, GEOLOCATION_API_URL } from '@/constants/constants'
  * @param {string[]} [fields]
  * @return {*}  {string}
  */
-export const getGeolocationApiUrl = (fields?: string[]): string => {
+export const getGeolocationApiEndpoint = (fields?: string[]): string => {
   const fieldsParam = fields && fields.length ? fields.join(',') : ''
-  return `${GEOLOCATION_API_URL}${GEOLOCATION_API_KEY}&fields=${fieldsParam}`
+  return `${GEOLOCATION_API_ENDPOINT}${GEOLOCATION_API_KEY}&fields=${fieldsParam}`
 }
